feat: Enable social encounter execution

+++ src/Explorer.API/Hubs/IEncounterExecutionClient.cs
@@ -3,7 +3,7 @@
     public interface IEncounterExecutionClient
     {
         Task StartSocialEncounter(long encounterId);
-        Task GetNumberOfActiveTourists(long numberOfActiveTours);
+        Task GetNumberOfActiveTourists(int numberOfActiveTours);
         Task CompleteSocialEncounter(long encounterId);
         Task Error(string message);
         Task AbandonSocialEncounter(long encounterId);
